requirements.txt
streamlit>=1.35.0
openai>=1.30.0
pymupdf>=1.24.0
pandas>=2.2.0
openpyxl>=3.1.0
requests>=2.32.0
python-dotenv>=1.0.0
